requirements.txt
# requirements.txt
sentence-transformers
faiss-cpu
numpy
pandas
scikit-learn
tk
